move "Look & Feel/Cover grid" to its own widget
